Add files via upload
QUESTION: Create a Snake, Water & Gun game in JavaScript. The game should ask you to enter S, W, or G. The computer should be able to randomly generate S, W, or G and declare a win or loss. Give the user the option to cancel the game if he doesn't want to play any longer.